Home

Some data set from Atomic Mass Data Center.

Usage

using AtomicMassEvaluation
He4 = get(AME2020, "He4")
binding_energy(He4)

\[28.2956624 \pm 8.0e-7\]

Types

AtomicMassEvaluation.Isotope — Type

struct Isotope
    Z::Int
    N::Int
end

Isotope struct, constructors as follows:

Isotope(iso::Isotope)
Isotope(Z::Integer, N::Integer)
Isotope(name::AbstractString) # parse the name of the isotope, e.g. `Ne20`

AtomicMassEvaluation.IsotopeData — Type

IsotopeData

Isotope is a struct that contains the information of an isotope. The information includes the decay mode, mass, binding energy, beta decay energy, and atomic mass.

Dateset

AtomicMassEvaluation.AME2020 — Constant

AME2020

the AME2020 dataset

Methods

AtomicMassEvaluation.getZ — Function

getZ(iso::Isotope)

get the proton number of the isotope

getZ(isodata::IsotopeData)

get the proton number of the IsotopeData

AtomicMassEvaluation.getN — Function

getN(iso::Isotope)

get the neutron number of the isotope

getN(isodata::IsotopeData)

get the neutron number of the IsotopeData

AtomicMassEvaluation.getA — Function

getA(iso::Isotope)

get the mass number of the isotope

getA(isodata::IsotopeData)

get the mass number of the IsotopeData

AtomicMassEvaluation.element — Function

element(iso::Isotope)

get the element symbol of the isotope

element(isodata::IsotopeData)

get the element symbol of the IsotopeData

AtomicMassEvaluation.name — Function

name(iso::Isotope)

get the name of the isotope

name(isodata::IsotopeData)

get the name of the IsotopeData

AtomicMassEvaluation.decaymode — Function

decaymode(isodata::IsotopeData)

get the decay mode of the IsotopeData

AtomicMassEvaluation.mass — Function

mass(isodata::IsotopeData)

get the mass of the IsotopeData

AtomicMassEvaluation.binding_energy — Function

binding_energy(isodata::IsotopeData)

get the binding energy of the IsotopeData

AtomicMassEvaluation.average_binding_energy — Function

average_binding_energy(isodata::IsotopeData)

get the average binding energy of the IsotopeData

AtomicMassEvaluation.beta_decay_energy — Function

beta_decay_energy(isodata::IsotopeData)

get the beta decay energy of the IsotopeData

AtomicMassEvaluation.atomic_mass — Function

atomic_mass(isodata::IsotopeData)

get the atomic mass of the IsotopeData

Base.get — Method

Base.get(data::Ref{Vector{Isotope}}, isoname::AbstractString)

get an IsotopeData from the dataset by name

Base.get — Method

Base.get(data::Ref{Vector{Isotope}}, isotope::Isotope)

get an IsotopeData from the dataset by isotope

Reference

Nuclear Data Services.

AME2020: Chinese Phys. C 45 030002 (2021), and Chinese Phys. C 45 030003 (2021).