Home

Some data set from Atomic Mass Data Center.

Usage

using AtomicMassEvaluation
He4 = get(AME2020, "He4")
binding_energy(He4)

\[28.2956624 \pm 8.0e-7\]

Types

AtomicMassEvaluation.Isotope — Type

Isotope

Isotope is a struct that contains the information of an isotope. The information includes the proton number, neutron number, element symbol, decay mode, mass, binding energy, beta decay energy, and atomic mass.

Dateset

AtomicMassEvaluation.AME2020 — Constant

AME2020

the AME2020 dataset

Methods

AtomicMassEvaluation.getZ — Function

getZ(iso::Isotope)

get the proton number of the isotope

AtomicMassEvaluation.getN — Function

getN(iso::Isotope)

get the neutron number of the isotope

AtomicMassEvaluation.getA — Function

getA(iso::Isotope)

get the mass number of the isotope

AtomicMassEvaluation.element — Function

element(iso::Isotope)

get the element symbol of the isotope

AtomicMassEvaluation.name — Function

name(iso::Isotope)

get the name of the isotope

AtomicMassEvaluation.decaymode — Function

decaymode(iso::Isotope)

get the decay mode of the isotope

AtomicMassEvaluation.mass — Function

mass(iso::Isotope)

get the mass of the isotope

AtomicMassEvaluation.binding_energy — Function

binding_energy(iso::Isotope)

get the binding energy of the isotope

AtomicMassEvaluation.average_binding_energy — Function

average_binding_energy(iso::Isotope)

get the average binding energy of the isotope

AtomicMassEvaluation.beta_decay_energy — Function

beta_decay_energy(iso::Isotope)

get the beta decay energy of the isotope

AtomicMassEvaluation.atomic_mass — Function

atomic_mass(iso::Isotope)

get the atomic mass of the isotope

Base.get — Method

Base.get(data::Ref{Vector{Isotope}}, isoname::AbstractString)

get an isotope from the dataset by name

Reference

Nuclear Data Services.

AME2020: Chinese Phys. C 45 030002 (2021), and Chinese Phys. C 45 030003 (2021).